avgiftspliktigInntektMnd er nullable. Endrer også til å være konsekvent med å bruke BigDecimal istedenfor å blande BigInteger in the miks. Da slipper man eventuelle feil som kommer av kovertering mellom de to.

+++ service/src/main/kotlin/no/nav/melosys/service/avgift/TrygdeavgiftsgrunnlagService.kt
@@ -92,8 +92,7 @@ class TrygdeavgiftsgrunnlagService(private val behandlingsresultatService: Behan
                 this.fomDato = fomDato
                 this.tomDato = tomDato
                 this.type = inntektskildeRequest.type
-                this.avgiftspliktigInntektMnd =
-                    Penger(inntektskildeRequest.avgiftspliktigInntektMnd.toBigDecimal())
+                this.avgiftspliktigInntektMnd = Penger(inntektskildeRequest.avgiftspliktigInntektMnd)
                 this.isArbeidsgiversavgiftBetalesTilSkatt = inntektskildeRequest.arbeidsgiversavgiftBetales
                 this.isTrygdeavgiftBetalesTilSkatt = this.utledTrygdeavgiftBetalesTilSkatt(request.skatteplikttype)
             }
